#10 Move tweets consumer to consumer package
